requirements.txt
scrapy
selenium
Django
djangorestframework
psycopg2-binary
Flask>=2.0.0
